+++ semcore/data-table/src/components/Body/Cell.types.ts
@@ -38,6 +38,7 @@ export type CellPropsInner<Data extends DataTableData, UniqKeyType> = {
   shadowVertical?: BodyPropsInner<Data, UniqKeyType>;
   lastLeftFixedIndex: number;
   firstRightFixedIndex: number;
+  withoutBorder?: boolean;
 };
 
 export type DataTableCellType = (<UniqKeyType, Tag extends Intergalactic.Tag = 'div'>(