+++ semcore/data-table/src/index.ts
@@ -8,6 +8,8 @@ import type {
   DataTableData,
   DataTableProps,
   DataTableChangeSort,
+  ColumnGroupConfig,
+  ColumnItemConfig,
 } from './components/DataTable/DataTable.types';
 
 const wrapDataTable = <PropsExtending extends {}>(
@@ -34,4 +36,6 @@ export type {
   DataTableData,
   DataTableProps,
   DataTableChangeSort,
+  ColumnGroupConfig,
+  ColumnItemConfig,
 };