+++ stories/components/mini-chart/tests/examples/trend-bars.tsx
@@ -2,24 +2,45 @@ import { Flex } from '@semcore/flex-box';
 import MiniChart from '@semcore/mini-chart';
 import React from 'react';
 
-const Demo = () => {
-  const data = [
-    { value: 10 },
-    { value: 20 },
-    { value: 50 },
-    { value: 80, color: 'chart-palette-order-1' },
-    { value: 45 },
-    { value: 66, color: 'chart-palette-order-5' },
-  ];
-
+type trendBarsProps = { loading?: boolean; animate?: boolean };
+const Demo = ({
+  loading,
+  animate,
+}: trendBarsProps) => {
   return (
-    <Flex id='mylabel'>
-      <MiniChart.TrendBar data={data} aria-hidden />
-      <MiniChart.TrendBar data={data} loading={true} aria-label='test' />
-      <MiniChart.TrendHistogram data={data} aria-labelledby='mylabel' />
-      <MiniChart.TrendHistogram data={data} loading={true} aria-describedby='mylabel' />
+    <Flex id='mylabel' direction='column'>
+      <Flex id='mylabel'>
+        <MiniChart.TrendBar
+          data={[
+            { value: 0.5 },
+            { value: 20 },
+            { value: 5 },
+            { value: 80, color: 'chart-palette-order-1' },
+            { value: 45 },
+            { value: 66, color: 'chart-palette-order-5' },
+            { value: 0, color: 'chart-palette-order-12' }]}
+          aria-hidden
+          loading={loading}
+          animate={animate}
+        />
+      </Flex>
+      <Flex id='mylabel'>
+        <MiniChart.TrendHistogram
+          data={[
+            { value: 0 },
+            { value: 21.5 },
+            { value: 50 },
+            { value: 80, color: 'chart-palette-order-1' },
+            { value: 45 },
+            { value: 66, color: 'chart-palette-order-5' }]}
+          aria-label='mylabel1'
+          loading={loading}
+          animate={animate}
+        />
+      </Flex>
     </Flex>
   );
 };
+export type { trendBarsProps };
 
 export default Demo;