+++ semcore/data-table/CHANGELOG.md
@@ -8,6 +8,10 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `variant` property. Adapts the table styling to different usage contexts, such as in a `Card`.
 
+### Fixed
+
+-  Bottom border for cell with empty data.
+
 ## [16.2.1] - 2025-09-05
 
 ### Changed