+++ semcore/data-table/CHANGELOG.md
@@ -10,7 +10,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Fixed
 
--  Bottom border for cell with empty data.
+- Bottom border for cell with empty data.
 
 ## [16.2.1] - 2025-09-05