+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.3.0] - 2025-09-12
+
+### Added
+
+- `variant` property. Adapts the table styling to different usage contexts, such as in a `Card`.
+- Export for `ColumnGroupConfig` and `ColumnItemConfig` types.
+
+### Fixed
+
+- Bottom border for cell with empty data.
+
 ## [16.2.1] - 2025-09-05
 
 ### Changed